requirements.txt
# --- CORE DATA SCIENCE ---
# Aligned to support both ML models and LLM integrations
numpy>=2.3.2
pandas>=2.3.1
scikit-learn>=1.7.2
xgboost>=3.1.3
scipy>=1.16.3
joblib>=1.5.2
matplotlib>=3.10.8
seaborn>=0.13.2
optuna>=4.7.0

# --- LLM & GENAI CORE (STRICT) ---
# Keeping these exactly as you have them to avoid import breakages
google-ai-generativelanguage==0.9.0
langchain-google-genai==2.1.12
pinecone==7.3.0
langchain-pinecone==0.2.13
openai==1.99.8
pydantic==2.11.7
pydantic-settings==2.10.1

# --- LANGCHAIN & AGENTIC STACK ---
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.79
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langsmith==0.4.41
langchain-text-splitters==0.3.9

# --- UTILITIES & INFRASTRUCTURE ---
python-dotenv==1.1.1
httpx==0.28.1
sqlalchemy==2.0.43
aiosqlite==0.21.0
tqdm==4.67.1
pyyaml==6.0.3
pillow==11.3.0
pypdf==6.0.0
beautifulsoup4==4.14.2

# --- DEPLOYMENT & TOOLS ---
streamlit==1.48.0
tavily-python
youtube-transcript-api==1.2.3